change on exec functions and modif parsing
