+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: atro <atro@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 07:56:05 by vgonnot           #+#    #+#             */
-/*   Updated: 2023/03/17 15:04:15 by atro             ###   ########.fr       */
+/*   Updated: 2023/03/21 10:01:49 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,8 @@ static void	exec_shell(char **env, t_env_main *main_env)
 		while (line != NULL)
 		{
 			add_history(line);
-			parsing(line, &all_cmd, env, main_env);
+			parsing(line, &all_cmd);
+			exec_cmd(&all_cmd, env, main_env);
 			free(line);
 			line = readline("prompt> ");
 		}
@@ -46,15 +47,12 @@ static void	exec_shell(char **env, t_env_main *main_env)
 int	main(int argc, char *argv[], char *env[])
 {
 	struct termios	termios_save;
-	t_env_main		*main_env;
+	t_env_main		main_env;
 
 	(void) argv;
 	check_param(argc);
-	main_env = malloc(sizeof(t_env_main));
-	if (!main_env)
-		exit(0);
-	main_env->tty = &termios_save;
-	init_main_env(main_env, env);
-	exec_shell(env, main_env);
+	main_env.tty = &termios_save;
+	init_main_env(&main_env, env);
+	exec_shell(env, &main_env);
 	return (0);
 }
